+++ usaspending_api/accounts/urls.py
@@ -7,7 +7,11 @@ tas_list = views.TreasuryAppropriationAccountViewSet.as_view(
     {'get': 'list', 'post': 'list'})
 tas_balances_list = views.TreasuryAppropriationAccountBalancesViewSet.as_view(
     {'get': 'list', 'post': 'list'})
+financial_accounts_by_award = views.FinancialAccountsByAwardListViewSet.as_view(
+    {'get': 'list', 'post': 'list'})
+
 urlpatterns = [
     url(r'^$', tas_balances_list),
-    url(r'^tas/', tas_list)
+    url(r'^tas/', tas_list),
+    url(r'^awards/', financial_accounts_by_award),
 ]
