+++ usaspending_api/accounts/views.py
@@ -2,6 +2,8 @@ from usaspending_api.accounts.models import TreasuryAppropriationAccount
 from usaspending_api.accounts.serializers import TreasuryAppropriationAccountSerializer
 from usaspending_api.accounts.models import AppropriationAccountBalances
 from usaspending_api.accounts.serializers import AppropriationAccountBalancesSerializer
+from usaspending_api.awards.serializers import FinancialAccountsByAwardsSerializer
+from usaspending_api.awards.models import FinancialAccountsByAwards
 from usaspending_api.common.mixins import FilterQuerysetMixin, ResponseMetadatasetMixin
 from usaspending_api.common.views import DetailViewSet
 from usaspending_api.common.mixins import SuperLoggingMixin
@@ -32,6 +34,26 @@ class TreasuryAppropriationAccountBalancesViewSet(SuperLoggingMixin,
 
     def get_queryset(self):
         queryset = AppropriationAccountBalances.objects.all()
+        queryset = self.serializer_class.setup_eager_loading(queryset)
+        filtered_queryset = self.filter_records(self.request, queryset=queryset)
+        ordered_queryset = self.order_records(self.request, queryset=filtered_queryset)
+        return ordered_queryset
+
+
+class FinancialAccountsByAwardListViewSet(
+        SuperLoggingMixin,
+        FilterQuerysetMixin,
+        ResponseMetadatasetMixin,
+        DetailViewSet):
+    """
+    Handles requests for financial account data grouped by award.
+    """
+
+    serializer_class = FinancialAccountsByAwardsSerializer
+
+    def get_queryset(self):
+        """Return the view's queryset."""
+        queryset = FinancialAccountsByAwards.objects.all()
         queryset = self.serializer_class.setup_eager_loading(queryset)
         filtered_queryset = self.filter_records(self.request, queryset=queryset)
         ordered_queryset = self.order_records(self.request, queryset=filtered_queryset)
