+++ usaspending_api/financial_activities/urls.py
@@ -5,10 +5,6 @@ from usaspending_api.financial_activities import views
 financial_accounts_by_program_object = views.FinancialAccountsByProgramActivityObjectClassListViewSet.as_view(
     {'get': 'list', 'post': 'list'})
 
-financial_accounts_by_award = views.FinancialAccountsByAwardListViewSet.as_view(
-    {'get': 'list', 'post': 'list'})
-
 urlpatterns = [
     url(r'^$', financial_accounts_by_program_object),
-    url(r'^award/', financial_accounts_by_award),
 ]
